pyproject.toml
[tool.poetry]
name = "backend_auth"
version = "0.1.0"
description = ""
authors = ["Pokutny Eugene <dead3@bk.ru>"]

[tool.poetry.dependencies]
python = "^3.9"
Flask = "^2.0.1"
alembic = "^1.6.5"
Flask-SQLAlchemy = "^2.5.1"
redis = "^3.5.3"
flask-marshmallow = "^0.14.0"
flasgger = "^0.9.5"
Flask-JWT-Extended = "^4.2.1"
gevent = "^21.1.2"
psycopg2-binary = "^2.9.1"
gunicorn = "^20.1.0"
requests = "^2.25.1"
pydantic = "^1.8.2"
Flask-Migrate = "^3.0.1"
Flask-Script = "^2.0.6"
Flask-Classful = "^0.14.2"
flask-expects-json = "^1.5.0"
email-validator = "^1.1.3"
marshmallow-sqlalchemy = "^0.26.1"
Flask-Security = "^3.0.0"
Flask-WTF = "^0.15.1"
Flask-Pydantic = "^0.8.0"
flask-pydantic-spec = "^0.1.4"
pydantic-sqlalchemy = "^0.0.9"

[tool.poetry.dev-dependencies]
Flask-DebugToolbar = "^0.11.0"
pytest-flask = "^1.2.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
